% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count}
\alias{count}
\title{counting function}
\usage{
count(variable, M)
}
\arguments{
\item{variable}{character or numeric, variable that should be counted}

\item{M}{data matrix in which to find the variable in}
}
\value{
numeric vector indicating the number of appearances of the variable of interest for each row of the provided data matrix
}
\description{
This function counts the appearance of a particular variable within each row of a data matrix
}

