% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_MMo.R
\name{create_MMo}
\alias{create_MMo}
\title{Calculating all pairwise assocations based on indicated metric}
\usage{
create_MMo(expMatrix, metricChoice)
}
\arguments{
\item{expMatrix}{numeric data matrix}

\item{metricChoice}{choice of metric, options include: Pearson Correlation "PE", Spearman Correlation "SP", Kendall Correlation "KE", Euclidean Distance "EU", Manhattan Distance "MA", and Canberra Distance "CA"}
}
\value{
ordered measure matrix, sorted from best to worst association measure
}
\description{
This function takes a data matrix and calculates all pairwise associations of the rows within the matrix. Association measures are sorted from best to worst, the ordered measure matrix is returned and written to a .RData file
}

