% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_smMMs.R
\name{create_smMMs}
\alias{create_smMMs}
\title{Internal function of makeAdjMatrices}
\usage{
create_smMMs(orderedMM, thresholds, metric)
}
\arguments{
\item{orderedMM}{measure matrix including all pairwise comparisions between vertex IDs based on a particular metric defining assosiction; association measured are sorted from best to worst}

\item{thresholds}{vector of numeric values indicating the thresholds on which to build the network models}

\item{metric}{charater (i.e. "SP", "PE", "EU", etc) indicating the metric used to define association between vertex IDs}
}
\value{
None
}
\description{
Splits the large ordered measure matrix into smaller matrices according to the indicated thresholds
}

