% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSWSFnetFromFile.R
\name{createSWSFnetFromFile}
\alias{createSWSFnetFromFile}
\title{Builds a small-world scale-free network}
\usage{
createSWSFnetFromFile(expMFile = NULL, metric = c("SP", "PE", "KE", "EU",
  "MA", "CA"), thresholds = NULL, MMoExists = FALSE)
}
\arguments{
\item{expMFile}{data file name (.txt) in tab-separated format, first column must be vertex ID (genes), first row are condition identifiers}

\item{metric}{charcter defining association between each vertex ID, correlation options: "SP" - Spearman Correlation, "PE" - Pearson Correlation, and "KE" - Kendall as define in cor{stats}; distances: "EU" - Euclidean, "MA" - Manhattan, and "CA" - canaberra as defined in dist{stats}}

\item{thresholds}{numeric vector representing a series of thresholds that the user can choose; this variable can only be assigned with the metric is one of the three correlation metrics}

\item{MMoExists}{wheter the measure matrix exists (from a previous run), set to FALSE, if set to TRUE it will load the measure matrix (file must be in the same directory as the run)}
}
\value{
None, output files are generated
}
\description{
This function creates a number of network models to determine the best threshold producing a small-world scale-free network. All network models are accessible as .RData files. The file winningThresh.RData includes the possible thresholds to construct a sw-sf network with the first entry representing the 'best' threshold for the model construction.
}

