% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataAssessToFile.R
\name{dataAssessToFile}
\alias{dataAssessToFile}
\title{Creates an output file providing general information about the data file and its analysis}
\usage{
dataAssessToFile(expMFile, metricUsed, CalculateThres, thresholds,
  winningThresh)
}
\arguments{
\item{expMFile}{filename containing data matrix}

\item{metricUsed}{metric being used to define association}

\item{CalculateThres}{whether user specified thresholds or not, boolean: TRUE or FALSE}

\item{thresholds}{list of thresholds considered for the network model generation}

\item{winningThresh}{'best' threshold that generates a scale-free, small-world network}
}
\value{
None; output file is generated
}
\description{
This function creates an output file including multiple infromation about the data at hand as well as steps taken in the analysis. This is an internal function of createSWSFnet
}

