% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataToVNs.R
\name{dataToVNs}
\alias{dataToVNs}
\title{Builds a small-world scale-free network and extracts VNs based on genes of interest}
\usage{
dataToVNs(dataFile, GoIFile = NULL, annoFile = NULL, metric = NULL,
  thresholds = NULL)
}
\arguments{
\item{dataFile}{data file name (.txt) in tab-separated format, first column must be vertex ID (genes), first row are condition identifiers}

\item{GoIFile}{a .txt file including vertex IDs of interest, IDs must be tab-separated}

\item{annoFile}{wheter there is an annoation file for the vertex IDs}

\item{metric}{charcter defining association between each vertex ID, correlation options: "SP" - Spearman Correlation, "PE" - Pearson Correlation, and "KE" - Kendall as define in cor{stats}; distances: "EU" - Euclidean, "MA" - Manhattan, and "CA" - canaberra as defined in dist{stats}}

\item{thresholds}{numeric vector representing a series of thresholds that the user can choose; this variable can only be assigned with the metric is one of the three correlation metrics}
}
\value{
None, output files are generated
}
\description{
This function calls createSWSFnetFromFile() and downstreamAnalysis(). Its a one strep function to create a scale-free small-world network, determines the 'best' network model, from which then vicinity networks (VNs) are extracted based on vertex IDs the user can upload. For more specifics refer to the two functions createSWSFnetFromFile() and  downstreamAnalysis()
}

