% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityM.R
\name{densityM}
\alias{densityM}
\title{Internal function of downstreamAnalysis, calculates the density of a network}
\usage{
densityM(FriendsM, FileOut = NULL, outReturn = TRUE)
}
\arguments{
\item{FriendsM}{subnetwork in matrix format; diagonal must be set to 1}

\item{FileOut}{name of file to write to if desired}

\item{outReturn}{whether to return the calculated density, boolean variable TRUE(default) or FALSE}
}
\value{
decimal between 0 and 1; zero indicating there are no edges in the network, and a network with density 1 is a clique (completely connected network).
}
\description{
The density of the passed network is calculated (number of edges in the graph versus number of all possible edges).
}

