% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityMatrix.R
\name{densityMatrix}
\alias{densityMatrix}
\title{Calculates the density of a provided network}
\usage{
densityMatrix(FriendsM, moreInfo = FALSE)
}
\arguments{
\item{FriendsM}{a network matrix with 0 and 1 indicating none- and existing links, respectively; diagonal must be set to 1}

\item{moreInfo}{whether to return more information, boolean variable TRUE or FALSE(default)}
}
\value{
decimal between 0 and 1; zero indicating there are no edges in the network, and a network with density 1 is a clique (completely connected network).
}
\description{
The density of the passed network is calculated (number of edges in the graph versus number of all possible edges).
}

