% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityNames.R
\name{densityNames}
\alias{densityNames}
\title{Calculates the density of a subnetwork based on vertex IDs}
\usage{
densityNames(Friends, adjMatrix, moreInfo = FALSE)
}
\arguments{
\item{Friends}{a list of names, which correspond to the associated network matrix's rownames}

\item{adjMatrix}{a network matrix with 0 and 1 indicating none- and existing links, respectively; diagonal must be set to 1}

\item{moreInfo}{whether to return more information, boolean variable TRUE or FALSE(default)}
}
\value{
decimal between 0 and 1; zero indicating there are no edges in the network, and a network with density 1 is a clique (completely connected network).
}
\description{
Calculates the density (number of edges in the graph versus number of all possible edges) of a subnetwork based on indicated rowname identifiers.
}

