% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstreamAnalysis.R
\name{downstreamAnalysis}
\alias{downstreamAnalysis}
\title{Automated subnetwork extraction}
\usage{
downstreamAnalysis(winnerT, metric, GoIFile, outFile, expMatrixFile,
  annoTFile = NULL, adjMatrix = NULL)
}
\arguments{
\item{winnerT}{threshold used to construct the network model}

\item{metric}{metrix used to construct the network model}

\item{GoIFile}{gene of interest (GoI) file, a txt file inlucuding vertex ids which are of particular interest}

\item{outFile}{indicate a filename to which to write the information to}

\item{expMatrixFile}{filename of original datafile (expression matrix)}

\item{annoTFile}{annotation file if there is one, first column must correspond to vertex IDs, first row must be column headers}

\item{adjMatrix}{custom adjacency matrix can be loaded if petal network construction was not used}
}
\value{
None, but multiple files are created
}
\description{
This function finds all maximal cliques of a subgroup of vertex ids, then the vicinity network (1-neighbor network) is identified for each maximal clique.
}

