% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findLargestCliques.R
\name{findLargestCliques}
\alias{findLargestCliques}
\title{Finds all largest cliques}
\usage{
findLargestCliques(adjMatrix)
}
\arguments{
\item{adjMatrix}{a network matrix with 0 and 1 indicating none- and existing links, respectively}
}
\value{
a numeric matrix; each column is a largest clique containing the vertex ids
}
\description{
This function calls the largest.cliques function from igraph version 0.7
}

