% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMaximalCliques.R
\name{findMaximalCliques}
\alias{findMaximalCliques}
\title{Finds all maximal cliques
 
This function calls the maximal.cliques function from igraph version 0.7}
\usage{
findMaximalCliques(adjMatrix)
}
\arguments{
\item{adjMatrix}{a network matrix with 0 and 1 indicating none- and existing links, respectively}
}
\value{
a list of containing numeric vetors of vertex ids; each list element is a clique
}
\description{
Finds all maximal cliques
 
This function calls the maximal.cliques function from igraph version 0.7
}

