% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findWinningThresh.R
\name{findWinningThresh}
\alias{findWinningThresh}
\title{Functions calculates best and alternative thresholds for a scale-free small-world network}
\usage{
findWinningThresh(threshFile, minR2 = 0.8, maxSlope = -1, minUsed = 40,
  minBigComp = 95)
}
\arguments{
\item{threshFile}{filename of file containing network statistics}

\item{minR2}{maximum acceptable slope of the linear regression of the log transformed degree distribution; default set to 0.8}

\item{maxSlope}{maximum acceptable slope of the linear regression of the log transformed degree distribution; default set to -1; slope should be -3, -1}

\item{minUsed}{minimum acceptable percentage of the original dataset used to build the network model; default is 40}

\item{minBigComp}{minimum acceptable size of the biggest component in percentage; default set to 95}
}
\value{
possibleThresh numeric vector presenting possible thresholds which construct scale-free small-world network models, the vector is sorted from best to worse
}
\description{
This function takes the network statistics file created by makeThresholdTable() and weights the parameters against each other to determine the best network that is scale-free and small-world and follows other important properties such as the consideration of the biggest component of a network model.
}

