% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateThresholds.R
\name{generateThresholds}
\alias{generateThresholds}
\title{Generates a series of thresholds based on the distribution of the numeric vector}
\usage{
generateThresholds(orderedV, adjMNum, decreasingOrder = TRUE)
}
\arguments{
\item{orderedV}{sorted numeric vector, for correlation and mutual infromation the vector should be in decending order, for distance the vector must be in increasing order (i.e. first entry should represent best association value)}

\item{adjMNum}{number indicating how many thresholds (i.e. network models) should be calculated}

\item{decreasingOrder}{wheter the orderedV is in decending order (TRUE) or increasing order (FALSE)}
}
\value{
numeric vector
}
\description{
This function generates a series of thresholds based on the passed numeric vector. These thresholds are used to generate a number of visible network models.
}

