% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVertexStats.R
\name{getVertexStats}
\alias{getVertexStats}
\title{Calculating all vertices' cluster coefficient and degree}
\usage{
getVertexStats(adjMatrix, totxtFile = FALSE)
}
\arguments{
\item{adjMatrix}{binary adjacency matrix, diagonal is assumed to be 1}

\item{totxtFile}{whether to right the table to a separate text file, boolean variable TRUE or FALSE(default)}
}
\value{
matrix with colnames "Index", "Gene","LocalCC","k"
}
\description{
Based on a network matrix, for each vertex of the network/row of the matrix, its cluster coeffient and connectivity is calculated and return in table format
}

