% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphHistQQ.R
\name{graphHistQQ}
\alias{graphHistQQ}
\title{Creates a tiff image of data histogram and the corresponding Q-Q plot}
\usage{
graphHistQQ(dataMatrix, fileName = "Data_Hist-QQ")
}
\arguments{
\item{dataMatrix}{numeric data matrix}

\item{fileName}{prefix of the desired output file, defaul is "Data_Hist-QQ"}
}
\value{
None; output file is generated
}
\description{
Creates a tiff files including the histogram of the data as well as the corresponding q-qplot see qqnorm{stats} for more information based on a numeric data matrix
}

