% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphHistQQFromFile.R
\name{graphHistQQFromFile}
\alias{graphHistQQFromFile}
\title{Creates a tiff image of data histogram and the corresponding Q-Q plot from a file}
\usage{
graphHistQQFromFile(inputFile, fileName = "Data_Hist-QQ")
}
\arguments{
\item{inputFile}{file name of a data matrix; file must be tab delimited, first line must be table header, first column must be identifiers, inner part of the matrix must be numeric}

\item{fileName}{prefix of the desired output file, defaul is "Data_Hist-QQ"}
}
\value{
None; output file is generated
}
\description{
Creates a tiff files including the histogram of the data as well as the corresponding q-qplot see qqnorm{stats} for more information. Input file is read in and the data matrix is stored as expM.RData.
}

