% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAdjMatrices.R
\name{makeAdjMatrices}
\alias{makeAdjMatrices}
\title{Function creating adjacency matrices for each threshold}
\usage{
makeAdjMatrices(orderedMMo, thresholds, metric)
}
\arguments{
\item{orderedMMo}{measure matrix including all pairwise comparisions between vertex IDs based on a particular metric defining assosiction; association measured are sorted from best to worst}

\item{thresholds}{thresholds defining similarity, list of numeric values}

\item{metric}{charater (i.e. "SP", "PE", "EU", etc) indicating the metric used to define association between vertex IDs}
}
\value{
None, output files are generated
}
\description{
For each threshold an binary adjacency matrix is constructed.
}

