% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCurrentSymAdjMatrix.R
\name{makeCurrentSymAdjMatrix}
\alias{makeCurrentSymAdjMatrix}
\title{Internal function of makeAdjMatrices}
\usage{
makeCurrentSymAdjMatrix(shm, currentThreshIndex, thresholds, edgeNum)
}
\arguments{
\item{shm}{metric prefix}

\item{currentThreshIndex}{threshold index that is currently considered}

\item{thresholds}{vector of numeric values}

\item{edgeNum}{numer of edges for the considered threshold}
}
\value{
edgeNum, the number of edges for the network model under considered threshold
}
\description{
Adjacency matricies are produced and checked for correctness
}

