% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCytoFile.R
\name{makeCytoFile}
\alias{makeCytoFile}
\title{Creates an network file for Cytoscape}
\usage{
makeCytoFile(threshold, metric, orderedMM)
}
\arguments{
\item{threshold}{numeric value indicating the threshold building a network model}

\item{metric}{charater (i.e. "SP", "PE", "EU", etc) indicating the metric used to define association between vertex IDs}

\item{orderedMM}{measure matrix including all pairwise comparisions between vertex IDs based on a particular metric defining assosiction; association measured are sorted from best to worst}
}
\value{
None; output file is generated
}
\description{
Creates a txt file which can be loaded as is into Cytoscape to plot the network based on the indicated threshold and metric.
}

