% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeThresholdTable.R
\name{makeThresholdTable}
\alias{makeThresholdTable}
\title{Calculatiing a varity of network statistics for a number of network models}
\usage{
makeThresholdTable(metricChoice, thresholds, numGenes,
  outputFile = "NetworkStats.txt")
}
\arguments{
\item{metricChoice}{chracter indicating metric choice, must be one of the following: "PE", "SP", "KE", "EU", "MA", "CA"}

\item{thresholds}{threholds considered to create the different network models}

\item{numGenes}{number of rows in original data matrix}

\item{outputFile}{file name set to "NetworkStats.txt", but can be adjusted}
}
\value{
none, output file is generated
}
\description{
This function calls all adjancency matrices which were build by makeAdjMatrices(), removed unconnected vertices from each network model and calculates network statistics for each model. Network statistics include: "R^2" and "slope/power" of the log-transformed connectrivity distribution to determine scale-freeness, "meanCC" and "meanPath" standing for the mean cluster coefficient and mean pathlength within the model to classify the network as small-world, "%used" indicating the percentage vertex IDs used from the original dataset, "%bigComp" the percentage of vertices included in the biggest component,"numComp" the number of components, "kmax" the maximal number of connections, "kmean" the average connectity of all verticies in the network, "kmedian" the median connectity of all verticies in the network, "diameter" of the network which is equal to the shortest longest path within the network model.
}

