% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeVNM.R
\name{makeVNM}
\alias{makeVNM}
\title{Creating a vicinity network martrix}
\usage{
makeVNM(currentSubgroup, adjMatrix)
}
\arguments{
\item{currentSubgroup}{a single node identifier or a string of node identifiers (rownames of adjMartrix)}

\item{adjMatrix}{boolean network matrix}
}
\value{
vicinity network of input node(s)
}
\description{
Creats a vicinity network based on one or multiple node. If created on multiple nodes the input nodes must be a clique in the network
}

