% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotExpProfiles.R
\name{plotExpProfiles}
\alias{plotExpProfiles}
\title{Graphing multiple vectors in one plot}
\usage{
plotExpProfiles(currentSubgroup, expMatrix, fileName, main = NULL,
  cl = NULL)
}
\arguments{
\item{currentSubgroup}{names of identifiers to be plotted}

\item{expMatrix}{data matrix which includes identifiers in its rownames}

\item{fileName}{file name to which the plot to save to, no file extension needed}

\item{main}{overall title for the plot}

\item{cl}{specification for the default plotting color, default set to rainbow (each vector has its own color, note differentiation between colors might be quite small)}
}
\value{
None
}
\description{
Plots multiple numeric vectors in rainbow color in not differently specified and saves the graph as a high resolution tiff
}

