% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProfiles.R
\name{plotProfiles}
\alias{plotProfiles}
\title{Internal function of plotExpProfiles}
\usage{
plotProfiles(expressionM, cond, header, farbe)
}
\arguments{
\item{expressionM}{datamatrix which to plot all indicies}

\item{cond}{colnames of the data matrix}

\item{header}{overall title for the plot}

\item{farbe}{specification for the default plotting color}
}
\value{
None
}
\description{
Creates canvas and scales the plot for plotExpProfiles().
}

