% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readinExpM.R
\name{readinExpM}
\alias{readinExpM}
\title{Custom file read function for a numeric data matrix}
\usage{
readinExpM(fileName)
}
\arguments{
\item{fileName}{file name which to read in, character sting with .txt}
}
\value{
readin numeric data matrix, rownames correspond to first column, colnames correspond to first row
}
\description{
Reads in the infromation from a file, file must be tab delimited, first line must be table header, first column must be identifiers, inner part of the matrix must be numeric
}

