% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readinTable.R
\name{readinTable}
\alias{readinTable}
\title{Custom file read function for a table format}
\usage{
readinTable(fileName)
}
\arguments{
\item{fileName}{file name which to read in, character sting with .txt}
}
\value{
matrix with no rownames
}
\description{
Reads in the information from a file in table format, file must be tab delimited, first line must be table header
}

