% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeToFile.R
\name{writeToFile}
\alias{writeToFile}
\title{Custom writing to file function}
\usage{
writeToFile(information, fileName, appending = TRUE)
}
\arguments{
\item{information}{material to be written to file, matrix or vector}

\item{fileName}{file name which to write to, charcter sting with .txt}

\item{appending}{whether to append to the specified file or overwrite, boolean variable TRUE(default) or FALSE}
}
\value{
None
}
\description{
This is a custom function to write information to file without row or column names
}

